Crate governance_os_voting

This crate is in charge of containing voting systems as supported by the Governance OS. They were made to integrate nicely with our own organizations pallet but should be reusable by other projects as well.

Structs

CoinBasedVotingParameters
Parameters for the coin based voting system.

ProposalMetadata
Metadata used and managed by all our voting systems. Need to be passed to their functions and persisted in storage.

Enums

VotingErrors

VotingSystems
This enum is in charge of representing and handling all existing voting systems.